refactor(storage): make segment common
