+++ src/storage/aol/mod.rs
@@ -1,2 +1,1 @@
 pub mod aol;
-pub mod segment;
