+++ src/storage/mod.rs
@@ -3,8 +3,9 @@ pub mod wal;
 
 use std::collections::HashMap;
 use std::fs::{read_dir, File, OpenOptions};
-use std::io::{self, Read, Write};
+use std::io::{self, Read, Seek, Write};
 use std::num::ParseIntError;
+use std::os::unix::fs::{FileExt, OpenOptionsExt};
 use std::path::{Path, PathBuf};
 
 use crc32fast::Hasher;
@@ -22,7 +23,7 @@ const BLOCK_SIZE: usize = 4096;
 /// used in the write-ahead log. The record header contains information
 /// about the type of the record, the length of the payload, and a checksum value.
 /// The constant is set to 8, reflecting the size of the record header structure.
-pub const RECORD_HEADER_SIZE: usize = 8;
+pub const WAL_RECORD_HEADER_SIZE: usize = 8;
 
 /// The magic value for identifying file headers.
 ///
@@ -120,7 +121,6 @@ impl CompressionLevel {
     }
 }
 
-
 /// A `Block` is an in-memory buffer that stores data before it is flushed to disk. It is used to
 /// batch writes to improve performance by reducing the number of individual disk writes. If the
 /// data to be written exceeds the `BLOCK_SIZE`, it will be split and flushed separately. The `Block`
@@ -168,7 +168,6 @@ impl<const BLOCK_SIZE: usize, const RECORD_SIZE: usize> Block<BLOCK_SIZE, RECORD
         self.written - self.flushed
     }
 }
-
 
 /// Represents options for configuring a segment in a write-ahead log.
 ///
@@ -225,18 +224,29 @@ pub struct Options {
     ///
     /// This is used by aol to cycle segments when the max file size is reached.
     max_file_size: u64,
+
+    /// A flag indicating whether the segment is a Write-Ahead Logging (WAL).
+    ///
+    /// If this flag is set to `true`, the segment's records will be encoded with the WAL header.
+    ///
+    /// If this flag is set to `false`, Write-Ahead Logging will be disabled, and records are encoded
+    /// without the WAL header.
+    ///
+    /// By default, this flag is set to `false`, indicating that Write-Ahead Logging is not enabled.
+    is_wal: bool,
 }
 
 impl Options {
     fn default() -> Self {
         Options {
-            dir_mode: Some(0o750),              // default directory mode
-            file_mode: Some(DEFAULT_FILE_MODE), // default file mode
-            compression_format: None,           // default compression format
-            compression_level: None,            // default compression level
-            metadata: None,                     // default metadata
-            extension: None,                    // default extension
-            max_file_size: DEFAULT_FILE_SIZE,   // default max file size (20mb)
+            dir_mode: Some(0o750),                                // default directory mode
+            file_mode: Some(DEFAULT_FILE_MODE),                   // default file mode
+            compression_format: Some(DEFAULT_COMPRESSION_FORMAT), // default compression format
+            compression_level: Some(DEFAULT_COMPRESSION_LEVEL),   // default compression level
+            metadata: None,                                       // default metadata
+            extension: None,                                      // default extension
+            max_file_size: DEFAULT_FILE_SIZE,                     // default max file size (20mb)
+            is_wal: false,
         }
     }
 
@@ -249,6 +259,7 @@ impl Options {
             metadata: None,
             extension: None,
             max_file_size: 0,
+            is_wal: false,
         }
     }
 
@@ -295,6 +306,11 @@ impl Options {
 
     fn with_dir_mode(mut self, dir_mode: u32) -> Self {
         self.dir_mode = Some(dir_mode);
+        self
+    }
+
+    fn with_wal(mut self) -> Self {
+        self.is_wal = true;
         self
     }
 }
@@ -500,8 +516,8 @@ fn encode_record(buf: &mut [u8], rec_len: usize, part: &[u8], i: usize) {
     buf[2..4].copy_from_slice(&len_part.to_be_bytes());
     buf[4..8].copy_from_slice(&crc.to_be_bytes());
 
-    // Copy the 'part' into the buffer starting from the RECORD_HEADER_SIZE offset
-    merge_slices(&mut buf[RECORD_HEADER_SIZE..], part); // Pass a mutable reference to buf
+    // Copy the 'part' into the buffer starting from the WAL_RECORD_HEADER_SIZE offset
+    merge_slices(&mut buf[WAL_RECORD_HEADER_SIZE..], part); // Pass a mutable reference to buf
 }
 
 // Reads a field from the given reader
@@ -534,10 +550,10 @@ fn write_file_header(file: &mut File, id: u64, opts: &Options) -> io::Result<usi
         id,
         opts.compression_format
             .as_ref()
-            .unwrap_or(&CompressionFormat::NoCompression),
+            .unwrap_or(&DEFAULT_COMPRESSION_FORMAT),
         opts.compression_level
             .as_ref()
-            .unwrap_or(&CompressionLevel::BestSpeed),
+            .unwrap_or(&DEFAULT_COMPRESSION_LEVEL),
     );
 
     if let Some(metadata) = &opts.metadata {
@@ -798,6 +814,323 @@ impl SegmentRef {
     }
 }
 
+/// Represents a segment in a write-ahead log.
+///
+/// A `Segment` represents a portion of the write-ahead log. It holds information about the file
+/// that stores the log entries, as well as details related to the segment's data and state.
+///
+/// A segment header is stored at the beginning of the segment file. It has the following format:
+///
+///
+///   File Header
+///
+///     0      1      2      3      4      5      6      7      8
+///     +------+------+------+------+------+------+------+------+
+///     | Magic                                                 |
+///     +------+------+------+------+------+------+------+------+
+///     | Version                                               |
+///     +------+------+------+------+------+------+------+------+
+///     | SegmentID                                             |
+///     +------+------+------+------+------+------+------+------+
+///     | Compression                                           |
+///     +------+------+------+------+------+------+------+------+
+///     | Compression Level                                     |
+///     +------+------+------+------+------+------+------+------+
+///     | Metadata                                              |
+///     .                                                       |
+///     .                                                       |
+///     .                                                       |
+///     +------+------+------+------+------+------+------+------+
+///
+pub(crate) struct Segment {
+    /// The unique identifier of the segment.
+    pub(crate) id: u64,
+
+    /// The path where the segment file is located.
+    pub(crate) file_path: PathBuf,
+
+    /// The active block for buffering data.
+    block: Block<BLOCK_SIZE, WAL_RECORD_HEADER_SIZE>,
+
+    /// The number of blocks that have been written to the segment.
+    written_blocks: usize,
+
+    /// The underlying file for storing the segment's data.
+    file: File,
+
+    /// The base offset of the file.
+    pub(crate) file_header_offset: u64,
+
+    /// The current offset within the file.
+    file_offset: u64,
+
+    /// A flag indicating whether the segment is closed or not.
+    closed: bool,
+
+    is_wal: bool,
+}
+
+impl Segment {
+    pub(crate) fn open(dir: &Path, id: u64, opts: &Options) -> io::Result<Self> {
+        // Ensure the options are valid
+        opts.validate()?;
+
+        // Build the file path using the segment name and extension
+        let extension = opts.extension.as_deref().unwrap_or("");
+        let file_name = segment_name(id, extension);
+        let file_path = dir.join(&file_name);
+        let file_path_exists = file_path.exists();
+        let file_path_is_file = file_path.is_file();
+
+        // Open the file with the specified options
+        let mut file = Self::open_file(&file_path, opts)?;
+
+        // Initialize the file header offset
+        let mut file_header_offset = 0;
+
+        // If the file already exists
+        if file_path_exists && file_path_is_file {
+            // Handle existing file
+            let header = read_file_header(&mut file)?;
+            validate_file_header(&header, id, opts)?;
+
+            file_header_offset += (4 + header.len());
+            let (index, _) = parse_segment_name(&file_name)?;
+            if index != id {
+                return Err(io::Error::new(
+                    io::ErrorKind::InvalidInput,
+                    "Invalid segment id",
+                ));
+            }
+            // TODO: take id etc from existing header
+        } else {
+            // Write new file header
+            let header_len = write_file_header(&mut file, id, opts)?;
+            file_header_offset += header_len;
+        }
+
+        // Seek to the end of the file to get the file offset
+        let file_offset = file.seek(io::SeekFrom::End(0))?;
+
+        // Initialize and return the Segment
+        Ok(Segment {
+            file,
+            file_header_offset: file_header_offset as u64,
+            file_offset: file_offset - file_header_offset as u64,
+            file_path: file_path,
+            id,
+            closed: false,
+            written_blocks: 0,
+            block: Block::new(),
+            is_wal: opts.is_wal,
+        })
+    }
+
+    fn open_file(file_path: &Path, opts: &Options) -> io::Result<File> {
+        let mut open_options = OpenOptions::new();
+        open_options.read(true).write(true);
+
+        if let Some(file_mode) = opts.file_mode {
+            open_options.mode(file_mode);
+        }
+
+        if !file_path.exists() {
+            open_options.create(true); // Create the file if it doesn't exist
+        }
+
+        open_options.open(file_path)
+    }
+
+    // Flushes the current block to disk.
+    // This method also synchronize file metadata to the filesystem
+    // hence it is a bit slower than fdatasync (sync_data).
+    pub(crate) fn sync(&mut self) -> io::Result<()> {
+        if self.closed {
+            return Err(io::Error::new(io::ErrorKind::Other, "Segment is closed"));
+        }
+
+        if self.block.written > 0 {
+            self.flush_block(true)?;
+        }
+        self.file.sync_all()
+    }
+
+    pub(crate) fn close(&mut self) -> io::Result<()> {
+        if self.closed {
+            return Err(io::Error::new(io::ErrorKind::Other, "Segment is closed"));
+        }
+        self.sync()?;
+        self.closed = true;
+        Ok(())
+    }
+
+    pub(crate) fn flush_block(&mut self, clear: bool) -> io::Result<()> {
+        let mut p = &mut self.block;
+        let clear = clear || p.is_full();
+
+        // No more data will fit into the block. Clear it and write to disk.
+        // The remaining data in the block will be zeroed out.
+        if clear {
+            p.written = BLOCK_SIZE; // Write till end of block.
+        }
+
+        let n = self.file.write(&p.buf[p.flushed..p.written])?;
+        p.flushed += n;
+        self.file_offset += n as u64;
+
+        // We flushed an entire block, prepare a new one.
+        if clear {
+            p.reset();
+            self.written_blocks += 1;
+        }
+
+        Ok(())
+    }
+
+    // Returns the current offset within the segment.
+    pub(crate) fn offset(&self) -> u64 {
+        self.file_offset + self.block.unwritten() as u64
+    }
+
+    /// Appends data to the segment.
+    ///
+    /// This method appends the given data to the segment. If the block is full, it is flushed
+    /// to disk. The data is written in chunks to the current block until the block is full.
+    ///
+    /// # Parameters
+    ///
+    /// - `buf`: The data to be appended.
+    ///
+    /// # Returns
+    ///
+    /// Returns the number of bytes successfully appended.
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if the segment is closed.
+    pub(crate) fn append(&mut self, mut rec: &[u8]) -> io::Result<(u64, usize)> {
+        // If the segment is closed, return an error
+        if self.closed {
+            return Err(io::Error::new(io::ErrorKind::Other, "Segment is closed"));
+        }
+
+        if rec.is_empty() {
+            return Err(io::Error::new(io::ErrorKind::Other, "buf is empty"));
+        }
+
+        let offset = self.offset();
+        let mut n = 0;
+        let mut i = 0;
+
+        while i == 0 || !rec.is_empty() {
+            n += self.write_record(&mut rec, i)?;
+            i += 1;
+        }
+
+        // Write the remaining data to the block
+        if self.block.written > 0 {
+            self.flush_block(false)?;
+        }
+
+        Ok((offset, n))
+    }
+
+    fn write_record(&mut self, rec: &mut &[u8], i: usize) -> io::Result<usize> {
+        let active_block = &mut self.block;
+        let remaining = std::cmp::min(active_block.remaining(), rec.len());
+        let partial_record = &rec[..remaining];
+        let buf = &mut active_block.buf[active_block.written..];
+
+        if self.is_wal {
+            encode_record(buf, rec.len(), partial_record, i);
+            active_block.written += partial_record.len() + WAL_RECORD_HEADER_SIZE;
+        } else {
+            merge_slices(buf, partial_record);
+            active_block.written += partial_record.len();
+        }
+
+        if active_block.is_full() {
+            self.flush_block(true)?;
+        }
+
+        *rec = &rec[remaining..];
+        Ok(remaining)
+    }
+
+    /// Reads data from the segment at the specified offset.
+    ///
+    /// This method reads data from the segment starting from the given offset. It reads
+    /// from the underlying file if the offset is beyond the current block's buffer. The
+    /// read data is then copied into the provided byte slice `bs`.
+    ///
+    /// # Parameters
+    ///
+    /// - `bs`: A byte slice to store the read data.
+    /// - `off`: The offset from which to start reading.
+    ///
+    /// # Returns
+    ///
+    /// Returns the number of bytes read and any encountered error.
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if the provided offset is negative or if there is an I/O error
+    /// during reading.
+    pub(crate) fn read_at(&self, bs: &mut [u8], off: u64) -> io::Result<usize> {
+        if self.closed {
+            return Err(io::Error::new(io::ErrorKind::Other, "Segment is closed"));
+        }
+
+        if off > self.offset() {
+            return Err(io::Error::new(
+                io::ErrorKind::Other,
+                "Offset beyond current position",
+            ));
+        }
+
+        // Calculate buffer offset
+        let mut boff = 0;
+
+        let mut n = 0;
+        if off < self.file_offset {
+            // Read from the file
+            n = self.file.read_at(bs, self.file_header_offset + off)?;
+        } else {
+            boff = (off - self.file_offset) as usize;
+        }
+
+        let pending = bs.len() - n;
+        if pending > 0 {
+            let available = self.block.unwritten() - boff;
+            let read_chunk_size = std::cmp::min(pending, available);
+
+            if read_chunk_size > 0 {
+                let buf = &self.block.buf
+                    [self.block.written + boff..self.block.written + boff + read_chunk_size];
+                merge_slices(bs, buf);
+            }
+
+            if read_chunk_size == pending {
+                return Ok(n);
+            } else {
+                return Err(io::Error::new(
+                    io::ErrorKind::UnexpectedEof,
+                    "Incomplete read",
+                ));
+            }
+        }
+
+        Ok(n)
+    }
+}
+
+impl Drop for Segment {
+    /// Attempt to fsync data on drop, in case we're running without sync.
+    fn drop(&mut self) {
+        self.close().ok();
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1032,18 +1365,18 @@ mod tests {
 
     #[test]
     fn test_remaining() {
-        let block: Block<4096, RECORD_HEADER_SIZE> = Block {
+        let block: Block<4096, WAL_RECORD_HEADER_SIZE> = Block {
             written: 100,
             flushed: 0,
             buf: [0; 4096],
         };
-        assert_eq!(block.remaining(), 3996 - RECORD_HEADER_SIZE);
+        assert_eq!(block.remaining(), 3996 - WAL_RECORD_HEADER_SIZE);
     }
 
     #[test]
     fn test_is_full() {
-        let block: Block<4096, RECORD_HEADER_SIZE> = Block {
-            written: 4096 - RECORD_HEADER_SIZE,
+        let block: Block<4096, WAL_RECORD_HEADER_SIZE> = Block {
+            written: 4096 - WAL_RECORD_HEADER_SIZE,
             flushed: 0,
             buf: [0; 4096],
         };
@@ -1052,7 +1385,7 @@ mod tests {
 
     #[test]
     fn test_reset() {
-        let mut block: Block<4096, RECORD_HEADER_SIZE> = Block {
+        let mut block: Block<4096, WAL_RECORD_HEADER_SIZE> = Block {
             written: 100,
             flushed: 0,
             buf: [1; 4096],
@@ -1096,4 +1429,373 @@ mod tests {
     //     assert_eq!(block.flushed, 0);
     // }
 
+    #[test]
+    fn test_aol_append() {
+        // Create a temporary directory
+        let temp_dir = TempDir::new("test").expect("should create temp dir");
+
+        // Create segment options and open a segment
+        let opts = Options::default();
+        let mut segment = Segment::open(&temp_dir.path(), 0, &opts).expect("should create segment");
+
+        // Test initial offset
+        let sz = segment.offset();
+        assert_eq!(0, sz);
+
+        // Test appending an empty buffer
+        let r = segment.append(&[]);
+        assert!(r.is_err());
+
+        // Test appending a non-empty buffer
+        let r = segment.append(&[0, 1, 2, 3]);
+        assert!(r.is_ok());
+        assert_eq!(4, r.unwrap().1);
+
+        // Test appending another buffer
+        let r = segment.append(&[4, 5, 6, 7, 8, 9, 10]);
+        assert!(r.is_ok());
+        assert_eq!(7, r.unwrap().1);
+
+        // Validate offset after appending
+        // 4 + 7 = 11
+        assert_eq!(segment.offset(), 11);
+
+        // Test syncing segment
+        let r = segment.sync();
+        assert!(r.is_ok());
+
+        // Validate offset after syncing
+        assert_eq!(segment.offset(), 4096);
+
+        // Test reading from segment
+        let mut bs = vec![0; 4];
+        let n = segment.read_at(&mut bs, 0).expect("should read");
+        assert_eq!(4, n);
+        assert_eq!(&[0, 1, 2, 3].to_vec(), &bs[..]);
+
+        // Test reading another portion of data from segment
+        let mut bs = vec![0; 7];
+        let n = segment.read_at(&mut bs, 4).expect("should read");
+        assert_eq!(7, n);
+        assert_eq!(&[4, 5, 6, 7, 8, 9, 10].to_vec(), &bs[..]);
+
+        // Test reading beyond segment's current size
+        let mut bs = vec![0; 15];
+        let r = segment.read_at(&mut bs, 4097);
+        assert!(r.is_err());
+
+        // Test appending another buffer after syncing
+        let r = segment.append(&[11, 12, 13, 14]);
+        assert!(r.is_ok());
+        assert_eq!(4, r.unwrap().1);
+
+        // Validate offset after appending
+        // 4096 + 4 = 4100
+        assert_eq!(segment.offset(), 4100);
+
+        // Test reading from segment after appending
+        let mut bs = vec![0; 4];
+        let n = segment.read_at(&mut bs, 4096).expect("should read");
+        assert_eq!(4, n);
+        assert_eq!(&[11, 12, 13, 14].to_vec(), &bs[..]);
+
+        // Test syncing segment again
+        let r = segment.sync();
+        assert!(r.is_ok());
+
+        // Validate offset after syncing again
+        assert_eq!(segment.offset(), 4096 * 2);
+
+        // Test closing segment
+        assert!(segment.close().is_ok());
+
+        // Cleanup: Drop the temp directory, which deletes its contents
+        drop(temp_dir);
+    }
+
+    #[test]
+    fn test_aol_reopen_empty_file() {
+        // Create a temporary directory
+        let temp_dir = TempDir::new("test").expect("should create temp dir");
+
+        // Create segment options and open a segment
+        let opts = Options::default();
+        let segment = Segment::open(&temp_dir.path(), 0, &opts).expect("should create segment");
+
+        // Test initial offset
+        assert_eq!(0, segment.offset());
+
+        drop(segment);
+
+        // Reopen segment should pass
+        let segment = Segment::open(&temp_dir.path(), 0, &opts).expect("should create segment");
+
+        // Test initial offset
+        assert_eq!(0, segment.offset());
+    }
+
+    #[test]
+    fn test_segment_reopen() {
+        // Create a temporary directory
+        let temp_dir = TempDir::new("test").expect("should create temp dir");
+
+        // Create segment options and open a segment
+        let opts = Options::default();
+        let mut segment = Segment::open(&temp_dir.path(), 0, &opts).expect("should create segment");
+
+        // Test initial offset
+        assert_eq!(0, segment.offset());
+
+        // Test appending a non-empty buffer
+        let r = segment.append(&[0, 1, 2, 3]);
+        assert!(r.is_ok());
+        assert_eq!(4, r.unwrap().1);
+
+        // Test appending another buffer
+        let r = segment.append(&[4, 5, 6, 7, 8, 9, 10]);
+        assert!(r.is_ok());
+        assert_eq!(7, r.unwrap().1);
+
+        // Validate offset after appending
+        // 4 + 7 = 11
+        assert_eq!(segment.offset(), 11);
+
+        // Test syncing segment
+        assert!(segment.sync().is_ok());
+
+        // Validate offset after syncing
+        assert_eq!(segment.offset(), 4096);
+
+        // Test closing segment
+        assert!(segment.close().is_ok());
+
+        drop(segment);
+
+        // Reopen segment
+        let mut segment = Segment::open(&temp_dir.path(), 0, &opts).expect("should create segment");
+
+        // Test initial offset
+        assert_eq!(segment.offset(), 4096);
+
+        // Test reading from segment
+        let mut bs = vec![0; 4];
+        let n = segment.read_at(&mut bs, 0).expect("should read");
+        assert_eq!(4, n);
+        assert_eq!(&[0, 1, 2, 3].to_vec(), &bs[..]);
+
+        // Test reading another portion of data from segment
+        let mut bs = vec![0; 7];
+        let n = segment.read_at(&mut bs, 4).expect("should read");
+        assert_eq!(7, n);
+        assert_eq!(&[4, 5, 6, 7, 8, 9, 10].to_vec(), &bs[..]);
+
+        // Test reading beyond segment's current size
+        let mut bs = vec![0; 15];
+        let r = segment.read_at(&mut bs, 4097);
+        assert!(r.is_err());
+
+        // Test appending another buffer after syncing
+        let r = segment.append(&[11, 12, 13, 14]);
+        assert!(r.is_ok());
+        assert_eq!(4, r.unwrap().1);
+
+        // Validate offset after appending
+        // 4096 + 4 = 4100
+        assert_eq!(segment.offset(), 4100);
+
+        // Test reading from segment after appending
+        let mut bs = vec![0; 4];
+        let n = segment.read_at(&mut bs, 4096).expect("should read");
+        assert_eq!(4, n);
+        assert_eq!(&[11, 12, 13, 14].to_vec(), &bs[..]);
+
+        // Test closing segment
+        assert!(segment.close().is_ok());
+
+        // Reopen segment
+        let segment = Segment::open(&temp_dir.path(), 0, &opts).expect("should create segment");
+        // Test initial offset
+        assert_eq!(segment.offset(), 4096 * 2);
+
+        // Cleanup: Drop the temp directory, which deletes its contents
+        drop(segment);
+        drop(temp_dir);
+    }
+
+    #[test]
+    fn test_segment_reopen_file() {
+        // Create a temporary directory
+        let temp_dir = TempDir::new("test").expect("should create temp dir");
+
+        // Create segment options and open a segment
+        let opts = Options::default();
+        let segment = Segment::open(&temp_dir.path(), 0, &opts).expect("should create segment");
+
+        // Test initial offset
+        assert_eq!(0, segment.offset());
+
+        drop(segment);
+
+        // Reopen segment should pass
+        let segment = Segment::open(&temp_dir.path(), 0, &opts).expect("should create segment");
+
+        // Test initial offset
+        assert_eq!(0, segment.offset());
+    }
+
+    #[test]
+    fn test_segment_corrupted_metadata() {
+        // Create a temporary directory
+        let temp_dir = TempDir::new("test").expect("should create temp dir");
+
+        // Create segment options
+        let opts = Options::default();
+
+        let mut segment = Segment::open(&temp_dir.path(), 0, &opts).expect("should create segment");
+
+        // Close the segment
+        segment.close().expect("should close segment");
+
+        // Corrupt the segment's metadata by overwriting the first few bytes
+        let segment_path = temp_dir.path().join("00000000000000000000");
+        let mut corrupted_data = vec![0; 4];
+
+        // Open the file for writing before writing to it
+        let mut corrupted_file = OpenOptions::new()
+            .read(true)
+            .write(true)
+            .open(&segment_path)
+            .expect("should open corrupted file");
+
+        corrupted_file
+            .write_all(&mut corrupted_data)
+            .expect("should write corrupted data to file");
+
+        // Attempt to reopen the segment with corrupted metadata
+        let reopened_segment = Segment::open(&temp_dir.path(), 0, &opts);
+        assert!(reopened_segment.is_err()); // Opening should fail due to corrupted metadata
+
+        // Cleanup: Drop the temp directory, which deletes its contents
+        drop(temp_dir);
+    }
+
+    #[test]
+    fn test_segment_closed_operations() {
+        // Create a temporary directory
+        let temp_dir = TempDir::new("test").expect("should create temp dir");
+
+        // Create segment options
+        let opts = Options::default();
+
+        // Create a new segment file and open it
+        let mut segment = Segment::open(&temp_dir.path(), 0, &opts).expect("should create segment");
+
+        // Close the segment
+        segment.close().expect("should close segment");
+
+        // Try to perform operations on the closed segment
+        let r = segment.append(&[0, 1, 2, 3]);
+        assert!(r.is_err()); // Appending should fail
+
+        let r = segment.sync();
+        assert!(r.is_err()); // Syncing should fail
+
+        let mut bs = vec![0; 12];
+        let n = segment.read_at(&mut bs, 0);
+        assert!(n.is_err()); // Reading should fail
+
+        // Reopen the closed segment
+        let mut segment = Segment::open(&temp_dir.path(), 0, &opts).expect("should reopen segment");
+
+        // Try to perform operations on the reopened segment
+        let r = segment.append(&[4, 5, 6, 7]);
+        assert!(r.is_ok()); // Appending should succeed on reopened segment
+
+        // Cleanup: Drop the temp directory, which deletes its contents
+        drop(temp_dir);
+    }
+
+    #[test]
+    fn test_wal_append() {
+        // Create a temporary directory
+        let temp_dir = TempDir::new("test").expect("should create temp dir");
+
+        // Create segment options and open a segment
+        let opts = Options::default().with_wal();
+        let mut segment = Segment::open(&temp_dir.path(), 0, &opts).expect("should create segment");
+
+        // Test initial offset
+        let sz = segment.offset();
+        assert_eq!(0, sz);
+
+        // Test appending an empty buffer
+        let r = segment.append(&[]);
+        assert!(r.is_err());
+
+        // Test appending a non-empty buffer
+        let r = segment.append(&[0, 1, 2, 3]);
+        assert!(r.is_ok());
+        assert_eq!(4, r.unwrap().1);
+
+        // Test appending another buffer
+        let r = segment.append(&[4, 5, 6, 7, 8, 9, 10]);
+        assert!(r.is_ok());
+        assert_eq!(7, r.unwrap().1);
+
+        // Validate offset after appending
+        // 8 + 4 + 8 + 7 = 27
+        assert_eq!(segment.offset(), 27);
+
+        // Test syncing segment
+        let r = segment.sync();
+        assert!(r.is_ok());
+
+        // Validate offset after syncing
+        assert_eq!(segment.offset(), 4096);
+
+        // Test reading from segment
+        let mut bs = vec![0; 12];
+        let n = segment.read_at(&mut bs, 0).expect("should read");
+        assert_eq!(12, n);
+        assert_eq!(&[0, 1, 2, 3].to_vec(), &bs[WAL_RECORD_HEADER_SIZE..]);
+
+        // Test reading another portion of data from segment
+        let mut bs = vec![0; 15];
+        let n = segment.read_at(&mut bs, 12).expect("should read");
+        assert_eq!(15, n);
+        assert_eq!(&[4, 5, 6, 7, 8, 9, 10].to_vec(), &bs[WAL_RECORD_HEADER_SIZE..]);
+
+        // Test reading beyond segment's current size
+        let mut bs = vec![0; 15];
+        let r = segment.read_at(&mut bs, 4097);
+        assert!(r.is_err());
+
+        // Test appending another buffer after syncing
+        let r = segment.append(&[11, 12, 13, 14]);
+        assert!(r.is_ok());
+        assert_eq!(4, r.unwrap().1);
+
+        // Validate offset after appending
+        // 4096 + 8 + 4 = 4108
+        assert_eq!(segment.offset(), 4108);
+
+        // Test reading from segment after appending
+        let mut bs = vec![0; 12];
+        let n = segment.read_at(&mut bs, 4096).expect("should read");
+        assert_eq!(12, n);
+        assert_eq!(&[11, 12, 13, 14].to_vec(), &bs[WAL_RECORD_HEADER_SIZE..]);
+
+        // Test syncing segment again
+        let r = segment.sync();
+        assert!(r.is_ok());
+
+        // Validate offset after syncing again
+        assert_eq!(segment.offset(), 4096 * 2);
+
+        // Test closing segment
+        assert!(segment.close().is_ok());
+
+        // Cleanup: Drop the temp directory, which deletes its contents
+        drop(temp_dir);
+    }
 }
