+++ src/storage/wal/mod.rs
@@ -1,5 +1,4 @@
 pub mod reader;
-pub mod segment;
 pub mod wal;
 
 use std::fmt;
